+++ c/src/046_goldbach.c
@@ -10,7 +10,7 @@
 static bool goldbach_represent (int, bool *);
 
 int main () {
-	size_t size = 1000;
+	int size = 1000;
 
 	int candidate = 9;
